refactor: Name 클래스 추가
 User의 이름을 저장하는 클래스
