+++ src/test/java/domain/user/NameTest.java
@@ -1,0 +1,24 @@
+package domain.user;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+class NameTest {
+
+    @Test
+    @DisplayName("빈 이름이 있는 경우 예외 처리")
+    void createPlayerWithEmptyName() {
+        assertThatThrownBy(() -> Name.of(""))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("빈 이름이 있습니다.");
+    }
+
+    @Test
+    void getName() {
+        Name name = Name.of("이름");
+        assertThat(name.getName()).isEqualTo("이름");
+    }
+}
